Refactor added movemtn file and extracted from movement method for each direction
