+++ src/key_events.py
@@ -5,6 +5,7 @@ from wall import Wall
 from hitbox import Hitbox
 from collision import game_over_voice, sword_slash_sound
 from hud import draw_hero_health
+from movement import movement
 from constants import (
     WIDTH,
     HEIGHT,
@@ -55,17 +56,17 @@ class Key_Events:
             self.game.attack_event('down', 0, 1)
 
 
-def movement(event, hero):
-
-    if event.key == pg.K_a:
-        hero.move(d_x=-1)
-        hero.load_movement_image(0)
-    if event.key == pg.K_d:
-        hero.move(d_x=1)
-        hero.load_movement_image(4)
-    if event.key == pg.K_w:
-        hero.move(d_y=-1)
-        hero.load_movement_image(8)
-    if event.key == pg.K_s:
-        hero.move(d_y=1)
-        hero.load_movement_image(12)
+# def movement(event, hero):
+#
+#     if event.key == pg.K_a:
+#         hero.move(d_x=-1)
+#         hero.load_movement_image(0)
+#     if event.key == pg.K_d:
+#         hero.move(d_x=1)
+#         hero.load_movement_image(4)
+#     if event.key == pg.K_w:
+#         hero.move(d_y=-1)
+#         hero.load_movement_image(8)
+#     if event.key == pg.K_s:
+#         hero.move(d_y=1)
+#         hero.load_movement_image(12)
